Move the retry logic to the SecurityKeyForm component
