+++ client/me/reauth-required/index.jsx
@@ -15,7 +15,6 @@ import Notice from 'calypso/components/notice';
 import WarningCard from 'calypso/components/warning-card';
 import { recordGoogleEvent } from 'calypso/state/analytics/actions';
 import { redirectToLogout } from 'calypso/state/current-user/actions';
-import { getCurrentUserId } from 'calypso/state/current-user/selectors';
 import SecurityKeyForm from './security-key-form';
 import TwoFactorActions from './two-factor-actions';
 import './style.scss';
@@ -134,12 +133,6 @@ class ReauthRequired extends Component {
 	preValidateAuthCode() {
 		return this.state.code.length && this.state.code.length > 5;
 	}
-
-	loginUserWithSecurityKey = () => {
-		return this.props.twoStepAuthorization.loginUserWithSecurityKey( {
-			user_id: this.props.currentUserId,
-		} );
-	};
 
 	renderFailedValidationMsg() {
 		if ( ! this.props.twoStepAuthorization.codeValidationFailed() ) {
@@ -286,10 +279,7 @@ class ReauthRequired extends Component {
 				isVisible={ this.props?.twoStepAuthorization.isReauthRequired() }
 			>
 				{ isSecurityKeySupported && twoFactorAuthType === 'webauthn' ? (
-					<SecurityKeyForm
-						loginUserWithSecurityKey={ this.loginUserWithSecurityKey }
-						onComplete={ this.refreshNonceOnFailure }
-					/>
+					<SecurityKeyForm twoStepAuthorization={ this.props.twoStepAuthorization } />
 				) : (
 					this.renderVerificationForm()
 				) }
@@ -319,18 +309,6 @@ class ReauthRequired extends Component {
 		);
 	}
 
-	refreshNonceOnFailure = ( error ) => {
-		const errors = [].slice.call( error?.data?.errors ?? [] );
-		const onErrorCallback = error?.onErrorCallback ?? ( () => {} );
-		if ( errors.some( ( e ) => e.code === 'invalid_two_step_nonce' ) ) {
-			this.props.twoStepAuthorization.fetch( () => {
-				this.loginUserWithSecurityKey().catch( () => onErrorCallback() );
-			} );
-		} else if ( error ) {
-			onErrorCallback();
-		}
-	};
-
 	handleAuthSwitch = ( authType ) => {
 		this.setState( { twoFactorAuthType: authType } );
 		if ( authType === 'sms' ) {
@@ -350,17 +328,12 @@ class ReauthRequired extends Component {
 }
 
 ReauthRequired.propTypes = {
-	currentUserId: PropTypes.number.isRequired,
+	twoStepAuthorization: PropTypes.object.isRequired,
 };
 
 /* eslint-enable jsx-a11y/no-autofocus, jsx-a11y/click-events-have-key-events, jsx-a11y/no-static-element-interactions, jsx-a11y/anchor-is-valid */
 
-export default connect(
-	( state ) => ( {
-		currentUserId: getCurrentUserId( state ),
-	} ),
-	{
-		redirectToLogout,
-		recordGoogleEvent,
-	}
-)( localize( ReauthRequired ) );
+export default connect( null, {
+	redirectToLogout,
+	recordGoogleEvent,
+} )( localize( ReauthRequired ) );
